Vim search highlight improvement
Previously highlight was disabled as soon as the text was entered into the search buffer. That was not optimal since I wished sometimes that search highlight was available, especially when tracking the origin of a variable.

The new configuration will not disable search highlight but a remap of ':noh' is done to <enter>. So whenever <enter> is pressed, search highligh is cleared