Move lsp stuff to unstaged as it is unconfirmed
Cosmetic changes for alacritty

Cosmetic changes for hyper-red.vim

Don't change color to cyan for last window in tmux. Too distracting.

Add fzf.vim - it's better to use this for vim, neovim and shell as it is common.
Instead of using multiple fuzzy finder for every application.

Move hyper-red.vim to vimcommon as it is common

Highlight linenumber in vim